Missing libs
During my tests i found out that in case of:
./crackmapexec.py 127.0.0.1 -u user -p pass --lsa 
Tool will (and it did) fail due to missing libraries. This applies to cases when the target is a legacy OS, such as WinXP or Win2k3.

I have fixed the issue by adding missing libraries. Not sure if there are any other missing libs in this file. ;]