Fix mssql check_if_admin function
The check_if_admin function from mssql.py takes an additional auth
parameter, that is actually not used. Other parts of the code are
calling the function without the parameter, which leads to an error when
enumerating mssql endpoints. By simply removing the parameter and fixing
the locations that use it, the issue gets resolved.